errors

+++ test/app.js
@@ -5,7 +5,8 @@ const fs = require('fs'),
   chai = require('chai'),
   chaiHttp = require('chai-http'),
   models = require('../app/models'),
-  dataCreation = require('../scripts/dataCreation');
+  dataCreation = require('../scripts/dataCreation'),
+  logger = require('../app//logger');
 
 chai.use(chaiHttp);
 
@@ -14,17 +15,20 @@ const getTablesQuery = `SELECT table_name FROM information_schema.tables WHERE t
 // THIS WORKS ONLY WITH POSTGRESQL
 beforeEach('drop tables, re-create them and populate sample data', done => {
   models.sequelize.query(getTablesQuery).then(tables => {
-    const tableExpression = tables
-      .map(table => {
-        return `"public"."${table[0]}"`;
-      })
-      .join(', ');
-    return models.sequelize
-      .query(`TRUNCATE TABLE ${tableExpression} RESTART IDENTITY`)
-      .then(() => {
-        return dataCreation.execute();
-      })
-      .then(() => done());
+    if (!tables.length) logger.info('Yuu have to run: NODE_ENV=testing npm run migrations');
+    else {
+      const tableExpression = tables
+        .map(table => {
+          return `"public"."${table[0]}"`;
+        })
+        .join(', ');
+      return models.sequelize
+        .query(`TRUNCATE TABLE ${tableExpression} RESTART IDENTITY`)
+        .then(() => {
+          return dataCreation.execute();
+        })
+        .then(() => done());
+    }
   });
 });
